refactor(vincent-sdk): Move jwt-specific 'auth' code into its own path, organize it and standardize its behaviour - VerifyJWT always throws errors with useful error messages, mimicking the DID-JWT package behaviour and using the same error message prefixes as it does - Creation and validation of tokens are separated since only verification will be relevant to SDK consumers (vincent app devs); creation is the domain of the Vincent dashboard frontend app - Utils JSDocs are marked as hidden unless they will be exported for consumers.
