--- packages/sdk/src/auth/index.ts
@@ -1,8 +1,0 @@
-export {
-  createPKPSigner,
-  createPKPSignedJWT,
-  verifyJWTSignature,
-  decodeJWT,
-} from './services/jwt-service';
-export { createJWTConfig } from './interfaces/create-jwt';
-export { DecodedJWT } from './interfaces/decode-jwt';
