+++ packages/sdk/src/jwt/core/utils/processJWTSignature.ts
@@ -1,5 +1,6 @@
-/**
- * Processes a JWT signature from base64url to binary
+/** Processes a JWT signature from base64url to binary
+ * @ignore
+ *
  * @param signature - The base64url encoded signature string
  * @returns A Uint8Array of the binary signature
  */
