+++ packages/sdk/src/jwt/index.ts
@@ -1,0 +1,17 @@
+import { createPKPSignedJWT } from './core/create';
+import { decodeJWT, verifyJWT } from './core/validate';
+import { isJWTExpired } from './core/utils';
+
+/** @namespace */
+export const jwt = {
+  /** @function
+   * @hidden
+   * */
+  create: createPKPSignedJWT,
+  /** @function */
+  decode: decodeJWT,
+  /** @function */
+  verify: verifyJWT,
+  /** @function */
+  isExpired: isJWTExpired,
+};
